State what each airframe can actually deliver
The x500 flying the Alia law is not only a naming problem, and the
airframe presets say why. Both declare the same operator envelope — the
same speed and rate limits, the same velocity caps — so the demand the
law shapes is right for both. What differs is the tilt the velocity loop
may command to reach it: 0.35 rad on the x500, 0.45 on the Alia. Level
flight at a tilt accelerates at g·tan(tilt), which puts a ceiling of
3.58 m/s² on one and 4.74 on the other.

Against those ceilings the shipped laws ask for more than either vehicle
has. Balanced asks 4.0, which the x500 cannot reach; Agile asks 6.5,
which neither can. Above the ceiling the demand slews faster than the
vehicle can follow, so the operator stops feeling the law and starts
feeling the airframe — and two modes that differ only above it feel the
same. That is the defect the direct family had, arriving from the other
direction: a difference stated in a file that nobody can fly.

The ceiling is recorded here as the airframe's own arithmetic rather
than as a number somebody chose, and three tests pin it: what each
ceiling is, which modes the x500 cannot reach, and that even the Alia
cannot fly its own Agile law. The laws themselves are left alone.
Changing a law that has been flown is a tuning decision and belongs to a
run that measures the result, not to the change that noticed the problem.

The Apple client's ownship work pushed two files past their size pins.
The position model split cleanly: reading this tablet's own sensors is a
different job from ranking which source the map believes, so the two
providers move to a file of their own, and the vehicle fix and the mark
it becomes move to another. HostLinkModel did not split — its relay
targets, controller loop and published state are one main-actor weave
held together by file-scoped `private`, exactly as the pin's own comment
warns, and an extension-file move only widens visibility. Its pin moves
deliberately instead, by the nine lines the feature added.
