+++ clients/apple/App/OwnshipDrawable.swift
@@ -1,0 +1,78 @@
+import Foundation
+import PilotageCore
+
+/// Where the vehicle says it is, as the link resolved it.
+///
+/// Metres per second here, knots at the display edge: the wire states SI and
+/// the conversion belongs where the reader is, not spread through the middle.
+struct VehicleFix: Equatable, Sendable {
+    let latitudeDegrees: Double
+    let longitudeDegrees: Double
+    /// Where the nose points, true, when the lane states an attitude.
+    let headingDegrees: Double?
+    /// Track over the ground, when the vehicle is moving fast enough for one.
+    let courseDegrees: Double?
+    /// Speed over the ground alongside `courseDegrees`.
+    let groundSpeedMetresPerSecond: Double?
+    /// Whether a simulator's oracle supplied it rather than the estimator.
+    let fromSimulator: Bool
+}
+
+extension VehicleFix {
+    /// Reads one link event, or nothing when it is not a vehicle fix.
+    ///
+    /// The event states six loose values; naming them is this type's job, not
+    /// the link's, so the link's switch stays a list of what arrived rather
+    /// than a place where meaning is assigned.
+    init?(_ event: LinkEvent) {
+        guard case let .vehicleFix(
+            latitudeDeg,
+            longitudeDeg,
+            headingDeg,
+            courseDeg,
+            groundSpeedMps,
+            fromSimulator
+        ) = event else { return nil }
+        self.init(
+            latitudeDegrees: latitudeDeg,
+            longitudeDegrees: longitudeDeg,
+            headingDegrees: headingDeg,
+            courseDegrees: courseDeg,
+            groundSpeedMetresPerSecond: groundSpeedMps,
+            fromSimulator: fromSimulator
+        )
+    }
+}
+
+extension OwnshipModel {
+    /// The resolved fix as the map draws it, or nothing when no source has one.
+    ///
+    /// Ground speed rides with the course and is converted at this edge, because the
+    /// display states knots and every lane below states metres per second.
+    ///
+    /// The heading is passed only when it is TRUE. A tablet's magnetometer reading is a
+    /// magnetic one, and a mark turned by it would point somewhere the vehicle is not;
+    /// the drawing side refuses a non-true reference, so this states the reference
+    /// rather than deciding for it.
+    func drawable(groundSpeedMetresPerSecond: Double?) -> DisplayOwnship? {
+        guard let fix else { return nil }
+        return DisplayOwnship(
+            coordinate: DisplayCoordinate(
+                latitudeDeg: fix.latitudeDegrees,
+                longitudeDeg: fix.longitudeDegrees
+            ),
+            courseDeg: fix.courseDegrees,
+            groundSpeedKt: groundSpeedMetresPerSecond.map { $0 * 1.943_844_49 },
+            headingDeg: heading.map(\.trueDegrees),
+            headingReference: heading.map {
+                switch $0.source {
+                case .deviceMagnetic: DisplayHeadingReference.magneticNorth
+                default: DisplayHeadingReference.trueNorth
+                }
+            },
+            altitudeFt: nil,
+            producerInstanceId: 0,
+            snapshotRevision: 0
+        )
+    }
+}
